Add Referee test - refactor from PlayfieldTest , reformat code
