+++ src/main/java/com/epam/prejap/tetris/Tetris.java
@@ -45,14 +45,14 @@ class Tetris {
     /**
      * Prepares the environment and launches the game.
      *
-     * @param args  array of strings input from the command line
-     *              <ul>
-     *                  <li>args[0] is dedicated to configuring custom navigation keys</li>
-     *                  <ul>
-     *                      <li>each key should be represented by a single character and separated by space</li>
-     *                      <li>input example: "q s d" -> none: q, left: s, right: d</li>
-     *                  </ul>
-     *              </ul>
+     * @param args array of strings input from the command line
+     *             <ul>
+     *                 <li>args[0] is dedicated to configuring custom navigation keys</li>
+     *                 <ul>
+     *                     <li>each key should be represented by a single character and separated by space</li>
+     *                     <li>input example: "q s d" -> none: q, left: s, right: d</li>
+     *                 </ul>
+     *             </ul>
      * @see CommandLineAnalyst#checkArgsForNavigationKeys(String)
      */
     public static void main(String[] args) {
