+++ src/test/resources/testng.xml
@@ -43,6 +43,7 @@
             <class name="com.epam.prejap.tetris.CommandLineAnalystTest"/>
             <class name="com.epam.prejap.tetris.game.PrinterTest"/>
             <class name="com.epam.prejap.tetris.game.TimerTest"/>
+            <class name="com.epam.prejap.tetris.game.RefereeTest"/>
         </classes>
     </test>
 </suite>
